cambio usuario

+++ microservice-usuario/src/main/java/com/microservice/usuario/repository/UsuarioRepository.java
@@ -5,7 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface UsuarioRepository extends JpaRepository<Usuario, Long>{
-    Optional<Usuario> findByUsername(String username); 
-
+public interface UsuarioRepository extends JpaRepository<Usuario, Integer> {
+    Optional<Usuario> findByUsername(String username);
 }
